Use react-native-aes for encryption
Temporarily move to react_native_crypto

Update bundle

Fix typo

Fix wrong names

Update bundle

Remove console.logs

Update naming for base64 instance

Add debugging to decrypt

Roll back decrypt debug

Debug generateUUID

Move generateUUIDSync
