+++ lib/crypto/abstract_crypto.js
@@ -13,15 +13,6 @@ export class SNAbstractCrypto {
     this.DefaultPBKDF2Length = 768;
   }
 
-  generateUUIDSync() {
-    return generateUUIDSync();
-  }
-
-  /* Can be overriden by native platforms to provide async implementation; defaults here to syncronous */
-  async generateUUID()  {
-    return generateUUIDSync();
-  }
-
   /* Constant-time string comparison */
   timingSafeEqual(a, b) {
     var strA = String(a);
@@ -81,6 +72,22 @@ export class SNAbstractCrypto {
     const messageData = CryptoJS.enc.Utf8.parse(message);
     const result = CryptoJS.HmacSHA256(messageData, keyData).toString();
     return result;
+  }
+
+  async hexStringToArrayBuffer(hex) {
+    for (var bytes = [], c = 0; c < hex.length; c += 2)
+    bytes.push(parseInt(hex.substr(c, 2), 16));
+    return new Uint8Array(bytes);
+  }
+
+  async base64ToArrayBuffer(base64) {
+    var binary_string = await this.base64Decode(base64);
+    var len = binary_string.length;
+    var bytes = new Uint8Array(len);
+    for(var i = 0; i < len; i++)        {
+      bytes[i] = binary_string.charCodeAt(i);
+    }
+    return bytes.buffer;
   }
 
   /** Generates two deterministic keys based on one input */
