+++ lib/crypto/cryptojs.js
@@ -1,7 +1,16 @@
 const CryptoJS = require("crypto-js");
+import { generateUUIDSync } from "@Lib/utils";
 import { SNAbstractCrypto } from './abstract_crypto';
 
 export class SNCryptoJS extends SNAbstractCrypto {
+
+  generateUUIDSync() {
+    return generateUUIDSync();
+  }
+
+  async generateUUID()  {
+    return generateUUIDSync();
+  }
 
   async pbkdf2(password, pw_salt, pw_cost, length) {
     const params = {
