+++ lib/crypto/react_native_crypto.js
@@ -5,19 +5,17 @@ export class SNReactNativeCrypto extends SNAbstractCrypto {
   // Called by React Native to dynamically set Aes and base64 module instances.
   setNativeModules({aes, base64}) {
     this.Aes = aes;
-    this.base64 = base64;
+    this.base64Instance = base64;
   }
 
   async generateUUID() {
-    return this.Aes.randomUuid().then((uuid) => {
-      return uuid.toLowerCase();
-    });
+    const uuid = await this.Aes.randomUuid();
+    return uuid.toLowerCase();
   }
 
   async pbkdf2(password, salt, cost, length) {
-    return this.Aes.pbkdf2(password, salt, cost, length).then(key => {
-      return key;
-    });
+    const key = await this.Aes.pbkdf2(password, salt, cost, length);
+    return key;
   }
 
   async generateRandomKey(length) {
@@ -29,11 +27,11 @@ export class SNReactNativeCrypto extends SNAbstractCrypto {
   }
 
   async base64(text) {
-    return this.base64.encode(text);
+    return this.base64Instance.encode(text);
   }
 
   async base64Decode(base64String) {
-    return this.base64.decode(base64String);
+    return this.base64Instance.decode(base64String);
   }
 
   async sha256(text) {
@@ -42,5 +40,15 @@ export class SNReactNativeCrypto extends SNAbstractCrypto {
 
   async hmac256(message, key) {
     return this.Aes.hmac256(message, key);
+  }
+
+  async aes256CbcDecrypt(ciphertext, keyData, ivData) {
+    return this.Aes.decrypt(ciphertext, keyData, ivData);
+  }
+
+  async aes256CbcEncrypt(text, keyData, ivData) {
+    const result = await this.Aes.encrypt(text, keyData, ivData);
+    console.log(result);
+    return result.cipher;
   }
 }
