+++ lib/crypto/webcrypto.js
@@ -1,5 +1,5 @@
 import { SNAbstractCrypto } from './abstract_crypto';
-import { getGlobalScope, getSubtleCrypto } from "@Lib/utils";
+import { getGlobalScope, getSubtleCrypto, generateUUIDSync } from "@Lib/utils";
 const subtleCrypto = getSubtleCrypto();
 
 export class SNWebCrypto extends SNAbstractCrypto {
@@ -7,6 +7,14 @@ export class SNWebCrypto extends SNAbstractCrypto {
   /**
   Public
   */
+
+  generateUUIDSync() {
+    return generateUUIDSync();
+  }
+
+  async generateUUID() {
+    return generateUUIDSync();
+  }
 
   async pbkdf2(password, pw_salt, pw_cost, length) {
     var key = await this.webCryptoImportKey(password, "PBKDF2", ["deriveBits"]);
@@ -161,22 +169,6 @@ export class SNWebCrypto extends SNAbstractCrypto {
       hexString += nextHexByte;
     }
     return hexString;
-  }
-
-  async hexStringToArrayBuffer(hex) {
-    for (var bytes = [], c = 0; c < hex.length; c += 2)
-    bytes.push(parseInt(hex.substr(c, 2), 16));
-    return new Uint8Array(bytes);
-  }
-
-  async base64ToArrayBuffer(base64) {
-    var binary_string = await this.base64Decode(base64);
-    var len = binary_string.length;
-    var bytes = new Uint8Array(len);
-    for(var i = 0; i < len; i++)        {
-      bytes[i] = binary_string.charCodeAt(i);
-    }
-    return bytes.buffer;
   }
 
   async arrayBufferToBase64(buffer) {
